fix(extensions): ensure __init__.py in extension dir is present
`gen_proto.sh` was deleting this file (or not recreating it) and if it
isn't present, then it isn't possible to grab the extension yaml files
in Python using `importlib.resources.files`.

This restores the missing __init__.py file and also adds a `touch` to
the `gen_proto.sh` script to make sure it sticks around.